Removed AvatarUtil
